use query from Stack O

+++ postgres_jun_2023/SQL_Usecase.groovy
@@ -7,6 +7,11 @@
 import groovy.sql.*
 import org.postgresql.*
 
+def emitCount = { table, count ->
+    String emitFormat = "# %-20s : %d"
+    println String.format(emitFormat, table, count)
+}
+
 def user = "postgres"
 def password = "swordfish"
 def host = "127.0.0.1"
@@ -15,37 +20,48 @@ def database = "sandbox"
 
 def sql = Sql.newInstance("jdbc:postgresql://${host}:${port}/${database}", user, password, "org.postgresql.Driver")
 
-// plan
-
+println "TRACER operation start"
 def planId = 150
-sql.execute("DELETE FROM bridge_reference_plan where plan_id = " + planId)
-sql.execute("DELETE FROM plan where id = " + planId)
 
-// new plan
-def insert
-insert = " INSERT INTO plan (id, name, status) VALUES (?,?,?); "
+sql.eachRow("SELECT COUNT(*) as c FROM bridge_reference_plan") { emitCount("brp", it.c) }
+sql.eachRow("SELECT COUNT(*) as c FROM bridge_reference_plan WHERE plan_id = ${planId}") { emitCount("brp ${planId}", it.c) }
 
-def name = 'target-plan'
-def status = 'active'
-sql.execute(insert, [planId, name, status]);
+def doSeed = false
 
-// associate one customer to all references
-def customerId = 301
-insert = """
-INSERT INTO bridge_customer_reference
-(customer_id, reference_id)
-SELECT ${customerId}, r.id FROM reference r
-"""
+if (doSeed) {
+    // delete plan
 
-sql.execute(insert);
+    sql.execute("DELETE FROM bridge_reference_plan where plan_id = " + planId)
+    sql.execute("DELETE FROM plan where id = " + planId)
+
+    // new plan
+    def insert
+    insert = " INSERT INTO plan (id, name, status) VALUES (?,?,?); "
+
+    def name = 'target-plan'
+    def status = 'active'
+    sql.execute(insert, [planId, name, status]);
+
+    // associate one customer to all references
+    def customerId = 301
+    insert = """
+    INSERT INTO bridge_customer_reference
+    (customer_id, reference_id)
+    SELECT ${customerId}, r.id FROM reference r
+    """
+
+    sql.execute(insert);
+}
+
+// sql.close();
+// System.exit(0);
 
 // write to bridge_reference_plan
 // for all customers, get all references and insert for plan
 
-def keys = []
-
 // mode 1
 /*
+def keys = []
 keys = sql.executeInsert("""
 INSERT INTO bridge_reference_plan
 (reference_id, plan_id)
@@ -55,37 +71,45 @@ JOIN reference r on r.id = brc.reference_id
 """)
 */
 
-// mode 2
-keys = sql.executeInsert("""
-INSERT INTO bridge_reference_plan
+sql.execute("""
+drop table if exists temp_brp;
+
+create temporary table temp_brp (
+	reference_id bigint,
+	plan_id bigint
+);
+
+-- insert all existing to temp
+-- todo: qualify by plan_id
+INSERT INTO temp_brp
 (reference_id, plan_id)
-SELECT DISTINCT r.id, ${planId} FROM customer c
+SELECT reference_id, plan_id FROM bridge_reference_plan
+WHERE plan_id = ${planId};
+
+-- insert all new to temp
+INSERT INTO temp_brp
+(reference_id, plan_id)
+SELECT r.id, ${planId} FROM customer c
 JOIN bridge_customer_reference brc on brc.customer_id = c.id
-JOIN reference r on r.id = brc.reference_id
+JOIN reference r on r.id = brc.reference_id;
+
+-- insert distinct from temp to original
+INSERT INTO bridge_reference_plan (reference_id, plan_id)
+     SELECT DISTINCT reference_id, plan_id
+     FROM temp_brp
+     WHERE NOT EXISTS (
+         SELECT 'X'
+         FROM bridge_reference_plan brp
+         WHERE
+             temp_brp.reference_id = brp.reference_id
+             AND temp_brp.plan_id = brp.plan_id
+     );
 """)
 
-println "TRACER insert # rows: " + keys.size()
+println "TRACER operation complete"
 
-println "TRACER results for plan " + planId
-
-sql.eachRow("""
-SELECT p.name as planName, r.name as refName FROM bridge_reference_plan brp
-JOIN reference r on r.id = brp.reference_id
-JOIN plan p on p.id = brp.plan_id
-WHERE brp.plan_id = ${planId}
-""") { row ->
-    println "plan: ${row.planName} reference: ${row.refName}"
-}
-
-println "TRACER cross-check: all references"
-
-sql.eachRow("""
-SELECT r.name as refName FROM customer c
-JOIN bridge_customer_reference bcr on bcr.customer_id = c.id
-JOIN reference r on r.id = bcr.reference_id
-""") { row ->
-    println "reference: ${row.refName}"
-}
+sql.eachRow("SELECT COUNT(*) as c FROM bridge_reference_plan") { emitCount("brp", it.c) }
+sql.eachRow("SELECT COUNT(*) as c FROM bridge_reference_plan WHERE plan_id = ${planId}") { emitCount("brp ${planId}", it.c) }
 
 sql.close()
